# Optimize `DbGte`