# Description of the Contribution

<!-- Describe the implementet feature or bugfix -->
Optimize some code in the creation of the tree structure in `DbGte`.
Not does not use a `BTreeMap` anymore, but a manually maintained sorted vector.
Lookups are done with `slice::binary_search`.

## Performance

Where the old (`GeneralizedTotalizer`) implementation was previously quite a
bit faster, for many inputs (here 100), the new (`DbGte`) implementation is now
faster.
For less inputs (50), the old implementation is still faster.
Also, when not encoding the entire tree, the old implementation tends to be
faster, since the new implementation computes all possible subsum values at
initialization, while the old one only does so when they are needed.

**Before** (at 3648bec)
```
test bench::pb_enc::dbgte_ub         ... bench:  88,486,775.90 ns/iter (+/- 20,036,162.14)
test bench::pb_enc::dbgte_ub_large   ... bench: 467,798,187.90 ns/iter (+/- 119,733,818.44)
test bench::pb_enc::gte_ub           ... bench:  55,420,211.10 ns/iter (+/- 4,144,646.04)
test bench::pb_enc::gte_ub_large     ... bench: 346,189,758.30 ns/iter (+/- 76,528,472.89)
```

**After** (at b72b57a)
```
test bench::pb_enc::dbgte_ub         ... bench:  62,062,022.40 ns/iter (+/- 19,406,803.27)
test bench::pb_enc::dbgte_ub_large   ... bench: 314,648,724.00 ns/iter (+/- 112,299,342.33)
test bench::pb_enc::gte_ub           ... bench:  55,835,263.10 ns/iter (+/- 10,072,418.51)
test bench::pb_enc::gte_ub_large     ... bench: 323,532,530.10 ns/iter (+/- 116,353,664.34)
```

<!-- If this PR resolves an open issue, please add a line "resolves #id" referencing the issue -->

# PR Checklist

- [x] I read and agree to [`CONTRIBUTING.md`](https://github.com/chrjabs/rustsat/blob/main/CONTRIBUTING.md)
- [x] I have formatted my code with `rustfmt` / `cargo fmt --all`
- [x] Commits are named following [conventional commits](https://www.conventionalcommits.org/en/v1.0.0/#summary)
- [x] I have added documentation for new features
- [x] The test suite still passes on this PR
- [x] I have added tests for new features / tests that would have caught the bug this PR fixes (please explain if not)
- [x] If this PR contains breaking changes, it is against the [`next-major`](https://github.com/chrjabs/rustsat/tree/next-major) branch, not against [`main`](https://github.com/chrjabs/rustsat/tree/main)